pyproject.toml
[project]
name = "goit-pythonweb-hw-08"
version = "0.1.0"
description = ""
authors = [
    {name = "sanyanut",email = "alexander.sorochinskiy@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standart] (>=0.136.1,<0.137.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic[email] (>=2.13.3,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic-extra-types (>=2.11.1,<3.0.0)",
    "phonenumbers (>=9.0.29,<10.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

package-mode = false